+++ BE/src/main/java/com/mse/datafabric/dataProducts/data/insights/DataProductInsightRepository.java
@@ -22,13 +22,13 @@ public class DataProductInsightRepository {
     @Autowired
     public GoogleMapsAPI googleMapsAPI;
 
-    public InsightFilterDTO getFilterById(int id, int filterId) {
+    public InsightFilterDTO getFilterById(long id, int filterId) {
         String SQL = "SELECT filter_column, type_id FROM insight_filters JOIN dataproducts ON dataproducts.id = insight_filters.dataproduct_id WHERE insight_filters.filter_id = ? AND dataproducts.id = ?";
         return jdbcTemplate.query(
                 SQL, new PreparedStatementSetter() {
                     public void setValues(PreparedStatement preparedStatement) throws SQLException {
                         preparedStatement.setInt(1, filterId);
-                        preparedStatement.setInt(2, id);
+                        preparedStatement.setLong(2, id);
                     }
                 },new ResultSetExtractor<>() {
                     public InsightFilterDTO extractData(ResultSet resultSet) throws SQLException,
@@ -43,13 +43,13 @@ public class DataProductInsightRepository {
                 }
         );
     }
-    public DataProductInsightDataDTO[] getDataProductInsights(int id){
+    public DataProductInsightDataDTO[] getDataProductInsights(long id){
         List<DataProductInsightDataDTO> dtoList = new ArrayList<>();
         String query = "SELECT * FROM dataproduct_insights JOIN dataproducts ON dataproducts.id = dataproduct_insights.dataproduct_id WHERE dataproducts.id = ? ORDER BY dataproduct_insights.id";
         jdbcTemplate.query(
                 query, new PreparedStatementSetter() {
                     public void setValues(PreparedStatement preparedStatement) throws SQLException {
-                        preparedStatement.setInt(1, id);
+                        preparedStatement.setLong(1, id);
                     }
                 },new ResultSetExtractor<>() {
                     public Object extractData(ResultSet resultSet) throws SQLException,
@@ -70,13 +70,13 @@ public class DataProductInsightRepository {
         );
         return dtoList.toArray(new DataProductInsightDataDTO[0]);
     }
-    public InsightFilterDTO[] getInsightFilters(int id){
+    public InsightFilterDTO[] getInsightFilters(long id){
         List<InsightFilterDTO> dtoList = new ArrayList<>();
         String query = "SELECT * FROM insight_filters JOIN dataproducts ON dataproducts.id = insight_filters.dataproduct_id WHERE dataproducts.id = ? ORDER BY insight_filters.type_id DESC";
         jdbcTemplate.query(
                 query, new PreparedStatementSetter() {
                     public void setValues(PreparedStatement preparedStatement) throws SQLException {
-                        preparedStatement.setInt(1, id);
+                        preparedStatement.setLong(1, id);
                     }
                 },new ResultSetExtractor<>() {
                     public Object extractData(ResultSet resultSet) throws SQLException,
@@ -114,7 +114,7 @@ public class DataProductInsightRepository {
             return null;
         }
     }
-    public boolean insertInsights(int id, DataProductInsightDataDTO[] dto){
+    public boolean insertInsights(long id, DataProductInsightDataDTO[] dto){
         final String STATEMENT = "INSERT INTO dataproduct_insights (type, display_name, unit, dataproduct_column, dataproduct_id) VALUES (?, ?, ?, ?, (SELECT id FROM dataproducts WHERE id = ?))";
         try {
             for(int i = 0;i< dto.length;i++){
@@ -125,7 +125,7 @@ public class DataProductInsightRepository {
                     ps.setString(2, dto[finalI].displayName);
                     ps.setString(3, dto[finalI].unit);
                     ps.setString(4, dto[finalI].dataProductColumn);
-                    ps.setInt(5, id);
+                    ps.setLong(5, id);
                     return ps;
                 });
             }
@@ -135,12 +135,12 @@ public class DataProductInsightRepository {
         }
         return true;
     }
-    public boolean deleteInsights(int id){
+    public boolean deleteInsights(long id){
         final String STATEMENT = "DELETE FROM dataproduct_insights USING dataproducts WHERE dataproduct_insights.dataproduct_id = dataproducts.id AND dataproducts.id = ?";
         try {
             jdbcTemplate.update(connection -> {
                 PreparedStatement ps = connection.prepareStatement(STATEMENT);
-                ps.setInt(1, id);
+                ps.setLong(1, id);
                 return ps;
             });
         }
@@ -149,13 +149,13 @@ public class DataProductInsightRepository {
         }
         return true;
     }
-    public DataProductInsightDataDTO[] getInsightsData(int id){
+    public DataProductInsightDataDTO[] getInsightsData(long id){
         final String STATEMENT = "SELECT type, display_name, unit, dataproduct_column, dataproduct_id FROM dataproduct_insights JOIN dataproducts ON dataproduct_insights.dataproduct_id = dataproducts.id WHERE dataproducts.id = ?";
         try {
             return jdbcTemplate.query(
                     STATEMENT, new PreparedStatementSetter() {
                         public void setValues(PreparedStatement ps) throws SQLException {
-                            ps.setInt(1, id);
+                            ps.setLong(1, id);
                         }
                     },new ResultSetExtractor<>() {
                         public DataProductInsightDataDTO[] extractData(ResultSet rs) throws SQLException,
@@ -178,13 +178,13 @@ public class DataProductInsightRepository {
             return null;
         }
     }
-    public InsightFilterDTO[] getInsightFiltersData(int id){
+    public InsightFilterDTO[] getInsightFiltersData(long id){
         final String STATEMENT = "SELECT type_id, filter_column, display_name, dataproduct_id FROM insight_filters JOIN dataproducts ON insight_filters.dataproduct_id = dataproducts.id WHERE dataproducts.id = ?";
         try {
             return jdbcTemplate.query(
                     STATEMENT, new PreparedStatementSetter() {
                         public void setValues(PreparedStatement ps) throws SQLException {
-                            ps.setInt(1, id);
+                            ps.setLong(1, id);
                         }
                     },new ResultSetExtractor<>() {
                         public InsightFilterDTO[] extractData(ResultSet rs) throws SQLException,
@@ -207,7 +207,7 @@ public class DataProductInsightRepository {
             return null;
         }
     }
-    public boolean insertInsightsFilter(int id, InsightFilterDTO[] dto){
+    public boolean insertInsightsFilter(long id, InsightFilterDTO[] dto){
         final String STATEMENT = "INSERT INTO insight_filters (type_id, filter_column, display_name, dataproduct_id) VALUES (?, ?, ?, (SELECT id FROM dataproducts WHERE id = ?))";
         try {
             for(int i = 0;i< dto.length;i++){
@@ -217,7 +217,7 @@ public class DataProductInsightRepository {
                     ps.setInt(1, dto[finalI].filterType);
                     ps.setString(2, dto[finalI].filterColumn);
                     ps.setString(3, dto[finalI].displayName);
-                    ps.setInt(4, id);
+                    ps.setLong(4, id);
                     return ps;
                 });
             }
@@ -227,12 +227,12 @@ public class DataProductInsightRepository {
         }
         return true;
     }
-    public boolean deleteInsightsFilter(int id){
+    public boolean deleteInsightsFilter(long id){
         final String STATEMENT = "DELETE FROM insight_filters USING dataproducts WHERE dataproducts.id = insight_filters.dataproduct_id AND dataproducts.id = ?";
         try {
             jdbcTemplate.update(connection -> {
                 PreparedStatement ps = connection.prepareStatement(STATEMENT);
-                ps.setInt(1, id);
+                ps.setLong(1, id);
                 return ps;
             });
         }
